安装卸载脚本可以用了

+++ symlink-dotfiles.sh
@@ -2,19 +2,26 @@
 # 更新配置文件软链接
 
 dotfiles=$(dirname $(readlink -f ${BASH_SOURCE[0]}))
+dotfiles_home=$dotfiles/home
 pushd $dotfiles > /dev/null
 
 link() {
     from="$1"
     to="$2"
-    echo "Linking '$from' to '$to'"
-    rm -f "$to"
+    if [[ -L $to ]]; then
+        echo "- remove symlink $to"
+        rm -f "$to"
+    elif [[ -f $to ]]; then
+        echo "- backup $to to $to.backup"
+        mv $to $to.backup
+    fi
+    echo "* Linking '$from' to '$to'"
     ln -s "$from" "$to"
 }
 
-for location in $(find files -maxdepth 1 | sed '1d'); do
-    file="${location##*/}"
-    link "$dotfiles/$location" "$HOME/.$file"
+for link_from in $(find $dotfiles_home -maxdepth 1 | sed '1d'); do
+    link_to="$HOME/.${link_from##*/}"
+    link "$link_from" "$link_to"
 done
 
 popd > /dev/null
